better comma control
